requirements.txt
aiogram==3.22.0
playwright
python-dotenv